requirements.txt
joblib
numpy
pyyaml
scipy
tqdm
gym
gym[atari]
